# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Bus Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  grade: string;
  guardianName: string;
  guardianPhone: string;
  biometricId: string;
  /** @nullable */
  busId?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  grade: string;
  /** @minLength 1 */
  guardianName: string;
  /** @minLength 1 */
  guardianPhone: string;
  /** @minLength 1 */
  biometricId: string;
  /** @nullable */
  busId?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  isActive?: boolean;
}

export interface StudentUpdate {
  /** @minLength 1 */
  name?: string;
  grade?: string;
  guardianName?: string;
  guardianPhone?: string;
  biometricId?: string;
  /** @nullable */
  busId?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  isActive?: boolean;
}

export interface Bus {
  id: number;
  busNumber: string;
  driverName: string;
  driverPhone: string;
  /** @nullable */
  routeId?: number | null;
  capacity?: number;
  isActive?: boolean;
  createdAt: string;
}

export interface BusInput {
  /** @minLength 1 */
  busNumber: string;
  /** @minLength 1 */
  driverName: string;
  /** @minLength 1 */
  driverPhone: string;
  /** @nullable */
  routeId?: number | null;
  capacity?: number;
  isActive?: boolean;
}

export interface BusUpdate {
  busNumber?: string;
  driverName?: string;
  driverPhone?: string;
  /** @nullable */
  routeId?: number | null;
  capacity?: number;
  isActive?: boolean;
}

export interface Route {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  totalStops?: number;
  /** @nullable */
  estimatedDurationMinutes?: number | null;
  createdAt: string;
}

export interface RouteInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  totalStops?: number;
  /** @nullable */
  estimatedDurationMinutes?: number | null;
}

export interface RouteUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  totalStops?: number;
  /** @nullable */
  estimatedDurationMinutes?: number | null;
}

export type TripStatus = typeof TripStatus[keyof typeof TripStatus];


export const TripStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface Trip {
  id: number;
  busId: number;
  /** @nullable */
  routeId?: number | null;
  status: TripStatus;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
  totalBoardings?: number;
  totalAlightings?: number;
  createdAt: string;
}

export interface TripInput {
  busId: number;
  /** @nullable */
  routeId?: number | null;
}

export type TripUpdateStatus = typeof TripUpdateStatus[keyof typeof TripUpdateStatus];


export const TripUpdateStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface TripUpdate {
  status?: TripUpdateStatus;
  /** @nullable */
  endedAt?: string | null;
}

export type ScanEventScanType = typeof ScanEventScanType[keyof typeof ScanEventScanType];


export const ScanEventScanType = {
  board: 'board',
  alight: 'alight',
} as const;

export interface ScanEvent {
  id: number;
  studentId: number;
  tripId: number;
  scanType: ScanEventScanType;
  scannedAt: string;
  /** @nullable */
  location?: string | null;
  smsSent?: boolean;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  guardianPhone?: string | null;
  /** @nullable */
  busNumber?: string | null;
}

export type ScanInputScanType = typeof ScanInputScanType[keyof typeof ScanInputScanType];


export const ScanInputScanType = {
  board: 'board',
  alight: 'alight',
} as const;

export interface ScanInput {
  /** @minLength 1 */
  biometricId: string;
  tripId: number;
  scanType: ScanInputScanType;
  /** @nullable */
  location?: string | null;
}

export interface SmsGateway {
  id: number;
  provider: string;
  apiUrl: string;
  apiKey?: string;
  senderId: string;
  isActive: boolean;
  updatedAt?: string;
}

export interface SmsGatewayInput {
  /** @minLength 1 */
  provider: string;
  /** @minLength 1 */
  apiUrl: string;
  /** @minLength 1 */
  apiKey: string;
  /** @minLength 1 */
  senderId: string;
  isActive?: boolean;
}

export type SmsLogStatus = typeof SmsLogStatus[keyof typeof SmsLogStatus];


export const SmsLogStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface SmsLog {
  id: number;
  studentId: number;
  guardianPhone: string;
  message: string;
  status: SmsLogStatus;
  /** @nullable */
  errorMessage?: string | null;
  sentAt: string;
  /** @nullable */
  studentName?: string | null;
}

export interface DashboardStats {
  totalStudents: number;
  totalBuses: number;
  totalRoutes?: number;
  activeTrips: number;
  todayScans: number;
  todayBoardings: number;
  todayAlightings: number;
  smsSentToday: number;
  smsFailedToday: number;
  studentsOnBus?: number;
}

export type ActivityItemScanType = typeof ActivityItemScanType[keyof typeof ActivityItemScanType];


export const ActivityItemScanType = {
  board: 'board',
  alight: 'alight',
} as const;

export interface ActivityItem {
  id: number;
  studentName: string;
  studentId: number;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  guardianName?: string | null;
  scanType: ActivityItemScanType;
  scannedAt: string;
  /** @nullable */
  busNumber: string | null;
  tripId: number;
  /** @nullable */
  location?: string | null;
  smsSent?: boolean;
  /** @nullable */
  travelDurationMinutes?: number | null;
}

export type ListStudentsParams = {
busId?: number;
search?: string;
};

export type ListTripsParams = {
busId?: number;
status?: ListTripsStatus;
};

export type ListTripsStatus = typeof ListTripsStatus[keyof typeof ListTripsStatus];


export const ListTripsStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export type ListScansParams = {
studentId?: number;
limit?: number;
};

export type ListSmsLogsParams = {
studentId?: number;
status?: ListSmsLogsStatus;
limit?: number;
};

export type ListSmsLogsStatus = typeof ListSmsLogsStatus[keyof typeof ListSmsLogsStatus];


export const ListSmsLogsStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};