# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School Bus Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buses:
    get:
      operationId: listBuses
      tags: [buses]
      summary: List all buses
      responses:
        "200":
          description: List of buses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bus"
    post:
      operationId: createBus
      tags: [buses]
      summary: Create a bus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
  /buses/{id}:
    get:
      operationId: getBus
      tags: [buses]
      summary: Get bus by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
        "404":
          description: Not found
    patch:
      operationId: updateBus
      tags: [buses]
      summary: Update bus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusUpdate"
      responses:
        "200":
          description: Updated bus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
    delete:
      operationId: deleteBus
      tags: [buses]
      summary: Delete bus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Routes ─────────────────────────────────────────────────────────────────
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes
      responses:
        "200":
          description: List of routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    get:
      operationId: getRoute
      tags: [routes]
      summary: Get route by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "404":
          description: Not found
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Trips ──────────────────────────────────────────────────────────────────
...
  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List recent scan events
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of scan events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanEvent"
    post:
      operationId: recordScan
      tags: [scans]
      summary: Record a biometric scan (board or alight)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Scan recorded and SMS sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanEvent"
        "400":
          description: Bad request

  # ── SMS Gateway ────────────────────────────────────────────────────────────
  /sms-gateway:
    get:
      operationId: getSmsGateway
      tags: [sms]
      summary: Get SMS gateway configuration
      responses:
        "200":
          description: SMS gateway config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsGateway"
        "404":
          description: Not configured
    put:
      operationId: upsertSmsGateway
      tags: [sms]
      summary: Create or update SMS gateway configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsGatewayInput"
      responses:
        "200":
          description: Updated SMS gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsGateway"
  /sms-logs:
    get:
      operationId: listSmsLogs
      tags: [sms]
      summary: List SMS logs
      parameters:
        - name: studentId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [sent, failed, pending]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of SMS logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmsLog"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Bus:
      type: object
      required: [id, busNumber, driverName, driverPhone, createdAt]
      properties:
        id:
          type: integer
        busNumber:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
        routeId:
          type: ["integer", "null"]
        capacity:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
    BusInput:
      type: object
      required: [busNumber, driverName, driverPhone]
      properties:
        busNumber:
          type: string
          minLength: 1
        driverName:
          type: string
          minLength: 1
        driverPhone:
          type: string
          minLength: 1
        routeId:
          type: ["integer", "null"]
        capacity:
          type: integer
        isActive:
          type: boolean
    BusUpdate:
      type: object
      properties:
        busNumber:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
        routeId:
          type: ["integer", "null"]
        capacity:
          type: integer
        isActive:
          type: boolean

    # ── Route ────────────────────────────────────────────────────────────────
    Route:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        totalStops:
          type: integer
        estimatedDurationMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        totalStops:
          type: integer
        estimatedDurationMinutes:
          type: ["integer", "null"]
    RouteUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        totalStops:
          type: integer
        estimatedDurationMinutes:
          type: ["integer", "null"]

    # ── Trip ─────────────────────────────────────────────────────────────────
...
    ScanEvent:
      type: object
      required: [id, studentId, tripId, scanType, scannedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        tripId:
          type: integer
        scanType:
          type: string
          enum: [board, alight]
        scannedAt:
          type: string
        location:
          type: ["string", "null"]
        smsSent:
          type: boolean
        studentName:
          type: ["string", "null"]
        guardianPhone:
          type: ["string", "null"]
        busNumber:
          type: ["string", "null"]
    ScanInput:
      type: object
      required: [biometricId, tripId, scanType]
      properties:
        biometricId:
          type: string
          minLength: 1
        tripId:
          type: integer
        scanType:
          type: string
          enum: [board, alight]
        location:
          type: ["string", "null"]

    # ── SmsGateway ────────────────────────────────────────────────────────────
    SmsGateway:
      type: object
      required: [id, provider, apiUrl, senderId, isActive]
      properties:
        id:
          type: integer
        provider:
          type: string
        apiUrl:
          type: string
        apiKey:
          type: string
        senderId:
          type: string
        isActive:
          type: boolean
        updatedAt:
          type: string
    SmsGatewayInput:
      type: object
      required: [provider, apiUrl, apiKey, senderId]
      properties:
        provider:
          type: string
          minLength: 1
        apiUrl:
          type: string
          minLength: 1
        apiKey:
          type: string
          minLength: 1
        senderId:
          type: string
          minLength: 1
        isActive:
          type: boolean

    # ── SmsLog ────────────────────────────────────────────────────────────────
    SmsLog:
      type: object
      required: [id, studentId, guardianPhone, message, status, sentAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        guardianPhone:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: string
        studentName:
          type: ["string", "null"]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...