# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Bus Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Route {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  totalStops?: number;
  /** @nullable */
  estimatedDurationMinutes?: number | null;
  createdAt: string;
}